test(connect): rewrite common.setup.js to ts
