+++ packages/connect/e2e/tests/device/authenticateDevice.test.ts
@@ -1,7 +1,7 @@
 import { DeviceAuthenticityConfig } from '@trezor/connect/src/data/deviceAuthenticityConfigTypes';
 import TrezorConnect from '../../../src';
 
-const { getController, setup, initTrezorConnect } = global.Trezor;
+import { getController, setup, initTrezorConnect } from '../../common.setup';
 
 const controller = getController();
