+++ packages/connect/e2e/tests/device/passphrase.test.ts
@@ -1,6 +1,6 @@
 import TrezorConnect from '../../../src';
 
-const { getController, setup, conditionalTest, initTrezorConnect } = global.Trezor;
+import { getController, setup, conditionalTest, initTrezorConnect } from '../../common.setup';
 
 const controller = getController();
