+++ packages/connect/e2e/tests/device/setBusy.test.ts
@@ -1,8 +1,8 @@
 import TrezorConnect from '../../../src';
 
-const { getController, setup, conditionalTest, initTrezorConnect } = global.Trezor;
+import { getController, setup, conditionalTest, initTrezorConnect } from '../../common.setup';
 
-const controller = getController('setBusy');
+const controller = getController();
 
 describe('TrezorConnect.setBusy', () => {
     beforeAll(async () => {
